+++ packages/std/src/addresses.rs
@@ -1,3 +1,5 @@
+#![allow(deprecated)]
+
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 use std::fmt;
@@ -86,6 +88,10 @@ impl From<Addr> for HumanAddr {
     }
 }
 
+#[deprecated(
+    since = "0.14.0",
+    note = "HumanAddr is not much more than an alias to String and it does not provide significant safety advantages. With CosmWasm 0.14, we now use String when there was HumanAddr before. There is also the new Addr, which holds a validated immutable human readable address."
+)]
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, Hash, JsonSchema)]
 pub struct HumanAddr(pub String);